Fix syntax and testing errors in test files and scripts
- Fix BigInt conversion in AccessControl.test.ts
- Fix assertion in TicketNFT.cov.test.ts for paused contract transfers
- Fix mint-tickets.ts structure to correctly mint tickets with proper parameters
- Improve import syntax in hardhat.config.ts

+++ test/unit/AccessControl.test.ts
@@ -40,13 +40,12 @@ describe('TicketNFT Access Control', () => {
 
   it('should prevent non-minters from minting tickets', async () => {
     const ticketMetadata = {
-  eventId: 1,
-  price: hardhat.ethers.parseEther('0.1'),
-  validFrom: BigInt(Math.floor(Date.now() / 1000) + 3600),
-  validUntil: BigInt(Math.floor(Date.now() / 1000) + 7200),
-  isTransferable: true,
-    
-};
+      eventId: 1,
+      price: hardhat.ethers.parseEther('0.1'),
+      validFrom: Math.floor(Date.now() / 1000) + 3600,
+      validUntil: Math.floor(Date.now() / 1000) + 7200,
+      isTransferable: true,
+    };
 
     await expect(
       ticketNFT.connect(nonMinter).mintTicket(nonMinter.address, 1, ticketMetadata.eventId, ticketMetadata.price, ticketMetadata.validFrom, ticketMetadata.validUntil, ticketMetadata.isTransferable)
@@ -55,13 +54,12 @@ describe('TicketNFT Access Control', () => {
 
   it('should allow minters to mint tickets', async () => {
     const ticketMetadata = {
-  eventId: 1,
-  price: hardhat.ethers.parseEther('0.1'),
-  validFrom: BigInt(Math.floor(Date.now() / 1000) + 3600),
-  validUntil: BigInt(Math.floor(Date.now() / 1000) + 7200),
-  isTransferable: true,
-    
-};
+      eventId: 1,
+      price: hardhat.ethers.parseEther('0.1'),
+      validFrom: Math.floor(Date.now() / 1000) + 3600,
+      validUntil: Math.floor(Date.now() / 1000) + 7200,
+      isTransferable: true,
+    };
 
     await expect(ticketNFT.connect(minter).mintTicket(minter.address, 1, ticketMetadata.eventId, ticketMetadata.price, ticketMetadata.validFrom, ticketMetadata.validUntil, ticketMetadata.isTransferable)).to.not.be
       .reverted;
